+++ src/components/organisms/LoginForm/style.ts
@@ -1,0 +1,9 @@
+import { Button } from 'components/atoms';
+import Styled from 'styled-components';
+import { palette } from 'styled-tools';
+
+export const LoginformWrap = Styled.div`
+  display : flex;
+  flex-direction : column;
+  align-items : center;
+`;
