+++ src/pages/Login/template/style.ts
@@ -9,6 +9,17 @@ height : 16px;
 object-fit : contain;
 `;
 
+export const LoginBtn = Styled(Button)`
+  margin-top : 14px;
+  border-radius: 4px;
+  background-image: linear-gradient(to right, ${palette('primary', 3)},${palette('primary', 0)} );
+  width: 406px;
+  height: 60px;
+  color : ${palette('grayscale', 0)};
+  font-size : 16px;
+  font-weight : bold;
+`;
+
 export const JoinBtn = Styled(Button)`
 margin-left : 15px;
 line-height : 1.5;
